requirements.txt
discord.py==2.2.2
aiohttp==3.8.4
dpy-http-server==0.1.0